qemu-yc: version 1:2.12.0-27

+++ monitor.c
@@ -2531,7 +2531,7 @@ int monitor_fdset_dup_fd_add(int64_t fdset_id, int dup_fd)
     return -1;
 }
 
-static int monitor_fdset_dup_fd_find_remove(int dup_fd, bool remove)
+static int64_t monitor_fdset_dup_fd_find_remove(int dup_fd, bool remove)
 {
     MonFdset *mon_fdset;
     MonFdsetFd *mon_fdset_fd_dup;
@@ -2554,7 +2554,7 @@ static int monitor_fdset_dup_fd_find_remove(int dup_fd, bool remove)
     return -1;
 }
 
-int monitor_fdset_dup_fd_find(int dup_fd)
+int64_t monitor_fdset_dup_fd_find(int dup_fd)
 {
     return monitor_fdset_dup_fd_find_remove(dup_fd, false);
 }
